+++ common/headers_test.go
@@ -26,7 +26,6 @@ const numHeadersToTest = 100
 func TestGetEthereumHeader(t *testing.T) {
 	rpcclient, _ := ethclient.Dial("https://eth.llamarpc.com")
 	header, _ := rpcclient.HeaderByNumber(context.Background(), big.NewInt(18495266))
-	fmt.Printf("Header: %v\n", header)
 	file, _ := os.Create("test_data/eth_header_18495266.json")
 	b, _ := header.MarshalJSON()
 	file.Write(b)
@@ -41,7 +40,7 @@ func TestTrueEthereumHeader(t *testing.T) {
 	headerBytes := make([]byte, 4096)
 	n, err := file.Read(headerBytes)
 	require.NoError(t, err)
-	fmt.Printf("Header bytes: %s\n", string(headerBytes[:n]))
+	//fmt.Printf("Header bytes: %s\n", string(headerBytes[:n]))
 	err = header.UnmarshalJSON(headerBytes[:n])
 	require.NoError(t, err)
 	var buffer bytes.Buffer
@@ -62,7 +61,7 @@ func TestFalseEthereumHeader(t *testing.T) {
 	headerBytes := make([]byte, 4096)
 	n, err := file.Read(headerBytes)
 	require.NoError(t, err)
-	fmt.Printf("Header bytes: %s\n", string(headerBytes[:n]))
+	//fmt.Printf("Header bytes: %s\n", string(headerBytes[:n]))
 	err = header.UnmarshalJSON(headerBytes[:n])
 	require.NoError(t, err)
 	hash := header.Hash()