+++ x/crosschain/genesis_test.go
@@ -13,7 +13,8 @@ import (
 
 func TestGenesis(t *testing.T) {
 	genesisState := types.GenesisState{
-		Params: types.DefaultParams(),
+		Params:         types.DefaultParams(),
+		ZetaAccounting: sample.ZetaAccounting(t, "sample"),
 		OutTxTrackerList: []types.OutTxTracker{
 			sample.OutTxTracker(t, "0"),
 			sample.OutTxTracker(t, "1"),