+++ changelog.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Features
+* [1395](https://github.com/zeta-chain/node/pull/1395) - Add state variable to track aborted zeta amount
 * [1387](https://github.com/zeta-chain/node/pull/1387) - Add HSM capability for zetaclient hot key
 
 ### Fixes